Moved param forward testing to wrapper folder
