+++ test/test_signature.py
@@ -10,16 +10,12 @@ import sys
 from inspect import Parameter, Signature, _empty
 from types import GetSetDescriptorType, ModuleType
 from typing import Any, Callable, ParamSpec, Type, TypeVar, cast, get_args, get_origin
-from unittest.mock import Mock, patch
-
-import networkx as nx
 import pytest
 
 from pyrigi import Framework
 from pyrigi.graph import Graph
 
 from test.wrapper._wrappers import _BadWrappers
-from test.wrapper._proxies import _BadWrappersBase
 
 P = ParamSpec("P")
 T = TypeVar("T")
@@ -266,198 +262,3 @@ def test_signature_graph(cls: Type):
         any_checked = True
 
     assert any_checked
-
-
-_MISSING = object()
-
-
-def _find_patch_target(method: Callable, wrapped_func: Callable) -> tuple:
-    """
-    Determine the module and attribute name to patch so that
-    ``patch.object(module, name)`` replaces ``wrapped_func`` at the
-    call-site used by ``method``.
-
-    Returns a ``(module, name)`` pair suitable for ``patch.object``.
-    """
-    method_module = sys.modules[method.__module__]
-
-    for attr_name, val in vars(method_module).items():
-        if val is wrapped_func:
-            return method_module, attr_name
-        if (
-            inspect.ismodule(val)
-            and getattr(val, wrapped_func.__name__, None) is wrapped_func
-        ):
-            return val, wrapped_func.__name__
-
-    # Fall back to the module where the proxy function is defined
-    return sys.modules[wrapped_func.__module__], wrapped_func.__name__
-
-
-def _assert_name_matches(cls: Type, attr_name: str, proxy_func_name: str) -> None:
-    assert attr_name == proxy_func_name, (
-        f"{cls.__name__}.{attr_name} is decorated with @copy_doc({proxy_func_name}) "
-        f"but the method name differs from the proxy function name"
-    )
-
-
-def _assert_params_forwarded(
-    cls: Type,
-    attr_name: str,
-    proxy_param_names: list,
-    mock_args: dict,
-    mock_func: Any,
-) -> None:
-    """
-    Assert that ``mock_func`` was called with the graph/framework instance
-    as the first positional argument and all entries of ``mock_args``
-    forwarded unchanged.
-
-    ``proxy_param_names`` is the ordered list of the proxy's parameter names
-    """
-    call_args = mock_func.call_args
-
-    # First positional arg must be the instance itself
-    first_arg = call_args[0][0]
-    if cls == _BadWrappers:
-        assert isinstance(
-            first_arg, _BadWrappersBase
-        ), f"{cls.__name__}.{attr_name} didn't pass "
-        "the _BadWrappers instance as first arg."
-    elif cls == Graph:
-        assert isinstance(
-            first_arg, nx.Graph
-        ), f"{cls.__name__}.{attr_name} didn't pass the Graph instance as first arg."
-    else:
-        assert isinstance(first_arg, Framework), (
-            f"{cls.__name__}.{attr_name} didn't pass "
-            f"the Framework instance as first arg."
-        )
-
-    # Every other parameter must be forwarded with the same value.
-    for param_name, expected_value in mock_args.items():
-        actual_value = call_args.kwargs.get(param_name, _MISSING)
-        if actual_value is _MISSING and len(call_args[0]) > 1:
-            try:
-                idx = proxy_param_names.index(param_name)
-                if idx < len(call_args[0]):
-                    actual_value = call_args[0][idx]
-            except ValueError:
-                pass
-
-        assert actual_value == expected_value, (
-            f"{cls.__name__}.{attr_name} didn't forward '{param_name}' correctly. "
-            f"Expected {expected_value}, got {actual_value}"
-        )
-
-    # Reverse check: no unexpected arguments must arrive at the proxy.
-    forwarded_names: set[str] = set(call_args.kwargs.keys())
-    for i, _ in enumerate(call_args[0][1:], start=1):
-        if i < len(proxy_param_names):
-            forwarded_names.add(proxy_param_names[i])
-
-    # Extra positionals beyond the proxy's named params are always unexpected.
-    extra_positional_count = max(0, len(call_args[0]) - len(proxy_param_names))
-    assert extra_positional_count == 0, (
-        f"{cls.__name__}.{attr_name} forwarded {extra_positional_count} unexpected "
-        f"positional argument(s) to the proxy"
-    )
-
-    unexpected = forwarded_names - set(mock_args.keys())
-    assert not unexpected, (
-        f"{cls.__name__}.{attr_name} forwarded unexpected argument(s) to the proxy: "
-        f"{unexpected}"
-    )
-
-
-@pytest.mark.parametrize(
-    ("cls", "expect_pass", "test_instance"),
-    [
-        (Graph, True, Graph([(0, 1), (1, 2)])),
-        (
-            Framework,
-            True,
-            Framework(
-                Graph([(0, 1), (1, 2)]),
-                {0: (0, 0), 1: (1, 0), 2: (0, 1)},
-            ),
-        ),
-        (_BadWrappers, False, _BadWrappers()),
-    ],
-)
-def test_wrapper_parameter_forwarding(cls, expect_pass, test_instance):
-    """
-    Test that all @copy_doc wrapper methods correctly forward parameters
-    to their underlying proxy functions, and that every class of
-    forwarding mistake is detected.
-    """
-    any_checked = False
-
-    for attr_name, method in inspect.getmembers(cls):
-        wrapped_func = getattr(method, "_wrapped_func", None)
-        if wrapped_func is None or not callable(method):
-            continue
-
-        sig = inspect.signature(wrapped_func)
-        proxy_param_names = list(sig.parameters.keys())
-
-        # Exclude VAR_KEYWORD/VAR_POSITIONAL; probe **kwargs with a sentinel key.
-        mock_args = {
-            name: Mock(name=name)
-            for name, param in list(sig.parameters.items())[1:]
-            if param.kind
-            not in (inspect.Parameter.VAR_KEYWORD, inspect.Parameter.VAR_POSITIONAL)
-        }
-        if any(
-            p.kind == inspect.Parameter.VAR_KEYWORD for p in sig.parameters.values()
-        ):
-            mock_args["_test_kwarg"] = Mock(name="_test_kwarg")
-
-        patch_module, patch_name = _find_patch_target(method, wrapped_func)
-
-        with patch.object(patch_module, patch_name) as mock_func:
-            try:
-                result = getattr(test_instance, attr_name)(**mock_args)
-            except Exception as e:
-                if expect_pass:
-                    pytest.fail(
-                        f"{cls.__name__}.{attr_name} raised an "
-                        f"unexpected exception while forwarding parameters — "
-                        f"the wrapper may be modifying an argument before "
-                        f"passing it: {type(e).__name__}: {e}"
-                    )
-
-            # If the wrapper returns a generator, consume it to trigger the call
-            if hasattr(result, "__iter__") and hasattr(result, "__next__"):
-                try:
-                    list(result)
-                except Exception:
-                    pass
-
-            if not expect_pass:
-                if not mock_func.called:
-                    continue  # proxy was never invoked
-                with pytest.raises(AssertionError) as error:
-                    _assert_name_matches(cls, attr_name, wrapped_func.__name__)
-                    _assert_params_forwarded(
-                        cls, attr_name, proxy_param_names, mock_args, mock_func
-                    )
-                print("Did not raise error: ", attr_name, error)
-            else:
-                assert mock_func.called, (
-                    f"{cls.__name__}.{attr_name} didn't call "
-                    f"{wrapped_func.__name__}"
-                )
-                _assert_name_matches(cls, attr_name, wrapped_func.__name__)
-                _assert_params_forwarded(
-                    cls, attr_name, proxy_param_names, mock_args, mock_func
-                )
-
-        any_checked = True
-
-    assert any_checked
-
-
-# def test_temp():
-#     with pytest.raises(TypeError):
-#         pass
